Change the order of the alias analysis passes to align with the opt pipeline of LLVM https://bugs.webkit.org/show_bug.cgi?id=133753
Patch by Juergen Ributzka <juergen@apple.com> on 2014-06-16
Reviewed by Geoffrey Garen.

The order in which the alias analysis passes are added affects also the
order in which they are utilized. Change the order to align with the
one use by LLVM itself. The last alias analysis pass added will be
evaluated first. With this change we first perform a basic alias
analysis and then use the type-based alias analysis (if required).

* ftl/FTLCompile.cpp:
(JSC::FTL::compile):

Identifier: 151895@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@170022 268f45cc-cd09-0410-ab3c-d52691b4dbfc