Snapshotting WKThumbnailViews should not tell Web processes backing unparented WKViews that they're in window https://bugs.webkit.org/show_bug.cgi?id=134134 <rdar://problem/17402119>
Reviewed by Anders Carlsson.

* UIProcess/API/mac/WKView.mm:
(-[WKView _setThumbnailView:]):
Re-evaluating our in-window state is only useful if we're going to use the live layer tree;
snapshots can be taken of unparented views with no trouble. Plus, doing so sets off a chain of
expensive things in the Web process which we can avoid if using snapshots.


Identifier: 152078@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@170217 268f45cc-cd09-0410-ab3c-d52691b4dbfc