REGRESSION(r170336): Crash in HTMLMediaElement::seekTimerFired() https://bugs.webkit.org/show_bug.cgi?id=134253
Rubber-stamped by Eric Carlson.

Now that seekWithTolerance() has been split up into an asyncronous section, the null-checks in seekWithTolerance()
need to be re-performed in seekTimerFired(), in case the m_player has been cleared between when the seek was initiated
and when the asynchronous section is performed.

* html/HTMLMediaElement.cpp:
(WebCore::HTMLMediaElement::seekTimerFired):

Identifier: 152225@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@170367 268f45cc-cd09-0410-ab3c-d52691b4dbfc