[ARMv7] Skip one particularly expensive fastFree() alignment check. <https://webkit.org/b/134402>
The size class alignment check in fastFree() was crazy expensive on ARMv7
due to its use of modulo. Disabling it shaves 470ms off of PLT main thread
time in the web process.

Note that this was very hard to identify in profiles due to TCO.

Reviewed by Gavin Barraclough.

* wtf/FastMalloc.cpp:
(WTF::TCMalloc_ThreadCache::CreateCacheIfNecessary):

Identifier: 152376@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@170536 268f45cc-cd09-0410-ab3c-d52691b4dbfc