[iOS][WK2] Do not put tap highlight on images that are not in links, and on applet/embed/object https://bugs.webkit.org/show_bug.cgi?id=134471 <rdar://problem/17513542>
Reviewed by Enrica Casucci.

UIKit has some special code to avoid tap highlight on certain type of object. This is not very flexible
since the page has no way of overriding that behavior.

This patch implement the same feature by changing the default stylesheet. The tap highlight is disabled
by default for those object, and enabled for images that are in a link.

* css/html.css:
(applet, embed, object, img):
(:-webkit-any-link img):


Identifier: 152453@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@170621 268f45cc-cd09-0410-ab3c-d52691b4dbfc