The OpenType MATH table is not read on Mac https://bugs.webkit.org/show_bug.cgi?id=133569
Reviewed by Chris Fleizach.

Source/WebCore:
An incorrect tag format is used to load the OpenType MATH table on Mac.
This commit makes the format consistent with what is used for other tables on Mac.
This will allow to load the MATH table and to use it for the MathML rendering.

No new tests. platform/graphics/opentype already contains tests for the OpenType MATH support.

* platform/graphics/opentype/OpenTypeMathData.cpp:

LayoutTests:
We update the references for some tests in LayoutTests/mathml/opentype/ and remove the failure from TestExpectations.
Other tests require a font with a MATH table to be installed on Mac bots and to be used by default (bug 133603).

* platform/mac-wk2/mathml/opentype/opentype-stretchy-expected.txt: Removed.
* platform/mac/TestExpectations: Enable mathml/opentype/opentype-stretchy-horizontal.html and mathml/opentype/opentype-stretchy.html.
* platform/mac/mathml/horizontal-LatinModern-expected.txt: Removed. This file should not have been in this directory.
* platform/mac/mathml/horizontal-LatinModern-munderover-expected.txt: Removed. This file should not have been in this directory.
* platform/mac/mathml/opentype-stretchy-horizontal-expected.txt: Removed. This file should not have been in this directory.
* platform/mac/mathml/opentype/opentype-stretchy-expected.txt: Add text reference.
* platform/mac/mathml/opentype/opentype-stretchy-horizontal-expected.txt: Add text reference.

Identifier: 152527@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@170699 268f45cc-cd09-0410-ab3c-d52691b4dbfc