ASSERT in SharedBuffer::maybeAppendDataArray() on MobileSafari launch https://bugs.webkit.org/show_bug.cgi?id=134812 <rdar://problem/17628434>
Reviewed by Joseph Pecoraro.

Fix bug in my fix in r170930. Initialize the badly named m_shouldUsePurgeableMemory field to
false. This field indicates when it is okay to use purgeable memory and is set to true once
the resource is finished loading. By setting it to true in the constructor we were creating
purgeable memory while the resource was still being loaded and this triggered the assert.

No new tests. Should be covered by existing tests.

* platform/cf/SharedBufferCF.cpp:
(WebCore::SharedBuffer::SharedBuffer):

Identifier: 152758@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@170975 268f45cc-cd09-0410-ab3c-d52691b4dbfc