[iOS][WK2] It should be safe to call WKContentViewInteraction's cleanupInteraction multiple times https://bugs.webkit.org/show_bug.cgi?id=134820
Patch by Benjamin Poulain <bpoulain@apple.com> on 2014-07-10
Reviewed by Andreas Kling.

If a view is destroyed just after a crash, "cleanupInteraction" is called twice: once on crash,
once on dealloc.

The code handling _interactionViewsContainerView is using KVO to monitor transform changes. It is not safe
to remove the observer if we are not already observing on that view.

To solve the problem, this patch makes the cleanup actually remove the view so that setup and cleanup
are completely symmetrical. If cleanup is called twice, the second time would not enter the branch because
the view is already nil.

* UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView setupInteraction]):
(-[WKContentView cleanupInteraction]):


Identifier: 152770@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@170988 268f45cc-cd09-0410-ab3c-d52691b4dbfc