[iOS Media] Start playback button should indicate when it can't play https://bugs.webkit.org/show_bug.cgi?id=134851
Reviewed by Sam Weinig.

If a video element has errors and cannot play, then the start
playback button should use the "broken" icon.

* Modules/mediacontrols/mediaControlsApple.js: Add a new "failed" class.
* Modules/mediacontrols/mediaControlsiOS.js:
(ControllerIOS.prototype.shouldHaveStartPlaybackButton): We need a playback button
even when there is an error.
(ControllerIOS.prototype.handleWirelessPickerButtonTouchStart): Don't offer Airplay
if we're in an error state.
(ControllerIOS.prototype.updateStatusDisplay): Set the "failed" class on
the playback buttons if we are in an error state.

Identifier: 152801@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@171028 268f45cc-cd09-0410-ab3c-d52691b4dbfc