Fixed position elements are misplaced when a WK1 view has contentInsets set https://bugs.webkit.org/show_bug.cgi?id=135031 -and corresponding- <rdar://problem/17682335>
Reviewed by Tim Horton.

[NSScrollView documentVisibleRect] includes content that is within the inset-area
of a view, but WebCore is interested in the content that is fully visible, so we
need to factor the inset sizes out of this rect.

Implement contract() to avoid the awkwardness of calling expand() with negative
values.
* platform/graphics/IntSize.h:
(WebCore::IntSize::contract):

Factor out insets
* platform/mac/ScrollViewMac.mm:
(WebCore::ScrollView::platformVisibleContentRect):
(WebCore::ScrollView::platformVisibleContentSize):



Identifier: 152977@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@171246 268f45cc-cd09-0410-ab3c-d52691b4dbfc